Time: 1763 ms (7.6%), Space: 8.8 MB (56.96%) - LeetHub

+++ 3133-minimum-array-end/3133-minimum-array-end.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long minEnd(int n, int x) {
-        int number=x;
+        long long number=x;
         for (int i=1; i<n;i++){
             number=(number+1) | x; 
         }
